Add files via upload
PYTHON BASIC ASSIGNMENT 3
SUBMITTED BY :SWAPNIL GOSAVI


1. Why are functions advantageous to have in your programs?
2. When does the code in a function run: when it's specified or when it's called?
3. What statement creates a function?
4. What is the difference between a function and a function call?
5. How many global scopes are there in a Python program? How many local scopes?
6. What happens to variables in a local scope when the function call returns?
7. What is the concept of a return value? Is it possible to have a return value in an expression?
8. If a function does not have a return statement, what is the return value of a call to that function?
9. How do you make a function variable refer to the global variable?
10. What is the data type of None?
11. What does the sentence import areallyourpetsnamederic do?
12. If you had a bacon() feature in a spam module, what would you call it after importing spam?
13. What can you do to save a programme from crashing if it encounters an error?
14. What is the purpose of the try clause? What is the purpose of the except clause?



Q.1) Why are functions advantageous to have in your programs?

Ans 1) Below are the advantages of functions in programming:-

Functions reduce the need for duplicate code. This makes programs shorter, easier to read, and easier to update. The main advantage of functions is code Reusability.

Increases program readability.

Divide a complex problem into simpler ones.

Reduces chances of error.

Modifying a program becomes easier by using function

Q.2) When does the code in a function run: when it's specified or when it's called?

Ans 2)
The code in a function executes when the function is called, not when the function is specified. When a function is "called" the program "leaves" the current section of code and begins to execute the first line inside the function. Example is mentioned below:

def my_function(fname):
  print(fname + " iNeuron")

my_function("Swapnil")
my_function("Sudhanshu")
my_function("Krish") 

Q 3) What statement creates a function?

The def statement defines a function

Syntax of Function:
def function_name(parameters): """doc string""" -----function body----- -----function body----- return value

def my_function(fname):
  print(fname + " iNeuron")

my_function("Swapnil")
my_function("Sudhanshu")
my_function("Krish") 

Q 4) What is the difference between a function and a function call?

Ans 4)
A function is a block of code that does a particular operation and returns a result. It usually accepts inputs as parameters and returns a result. The parameters are not mandatory. A function call is the code used to pass control to a function

Q 5) How many global scopes are there in a Python program? How many local scopes?

There is one global scope, and a local scope is created whenever a function is called. A variable created inside a function belongs to the local scope of that function, and can only be used inside that function whereas A variable created in the main body of the Python code is a global variable and belongs to the global scope.

#Example of Local Scope
def myfunc():
  x = 300
  def myinnerfunc():
    print(x)
  myinnerfunc()

myfunc() 

#Example of Global Scope
x = 300

def myfunc():
  print(x)

myfunc()

print(x) 

Q 6) What happens to variables in a local scope when the function call returns?

Ans 6) When a function returns, the local scope is destroyed, and all the variables in it are forgotten. A local variable becomes undefined after the function call completes

Q 7) What is the concept of a return value? Is it possible to have a return value in an expression?

The Python return statement is a key component of functions and methods. We can use the return statement to make functions send Python objects back to the caller code. These objects are known as the function’s return value. A return value is the value that a function call evaluates to. Like any value, a return value can be used as part of an expression.

Q 8) If a function does not have a return statement, what is the return value of a call to that function?

If there is no return statement for a function, its return value is None. The function always returns None if explicit return is not written.

Q 9) How do you make a function variable refer to the global variable?

Ans 9) A global statement will force a variable in a function to refer to the global variable. If you want to refer to a global variable in a function, you can use the global keyword to declare which variables are global.

Q 10) What is the data type of None?

Ans 10) The data type of None is NoneType.

Q 11) What does the sentence import areallyourpetsnamederic do?

Ans 11) That import statement imports a module named areallyourpetsnamederic.

Q 12) If you had a bacon() feature in a spam module, what would you call it after importing spam?

This function can be called with spam.bacon().

Q 13) What can you do to save a programme from crashing if it encounters an error?

We can place the line of code that might cause an error in a try clause and use except block to handle the error as mentioned below:

try:
  print(x)
except:
  print("An exception occurred") 

Q 14) What is the purpose of the try clause? What is the purpose of the except clause?

x = 700
try:
  print(x)
except NameError:
  print("Variable x is not defined")
except:
  print("Something else went wrong") 

# THANK YOU 
SWAPNIL GOSAVI